added achievement functionality and people api for leaderboard

+++ fittness-app/parent-module/admin-coach-service/src/main/java/tudor/work/dto/AchievementRequestDto.java
@@ -21,7 +21,7 @@ public class AchievementRequestDto {
 
     private String description;
 
-    private Integer numberOfPoints;
+    private Double numberOfPoints;
 
     private MultipartFile achievementPicture;
